requirements.txt
opencv-python==4.10.0.84
numpy==1.26.4
PyYAML==6.0.2
moviepy==1.0.3
google-api-python-client==2.151.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
Flask==3.0.3
requests==2.32.3
